+++ R/xmpdf-package.R
@@ -1,0 +1,7 @@
+#' @section Package options:
+#'  The following `xmpdf` option may be set globally via [base::options()]:
+#'   \describe{
+#'     \item{xmpdf_default_lang}{Set new default `default_lang` argument value for [as_lang_alt()].}
+#'   }
+#' @name xmpdf
+"_PACKAGE"