+++ tests/testthat/test-xmp.R
@@ -5,10 +5,8 @@ test_that("get_xmp() / set_xmp()", {
     skip_if_not(supports_get_xmp() && supports_set_xmp())
     f <- tempfile(fileext = ".pdf")
     on.exit(unlink(f))
-    pdf(f, onefile = TRUE)
+    pdf(f)
     grid.text("Page 1")
-    grid.newpage()
-    grid.text("Page 2")
     invisible(dev.off())
 
     expect_snapshot(print(xmp()))
@@ -37,7 +35,7 @@ test_that("get_xmp() / set_xmp()", {
     set_xmp(x, f)
     x2 <- get_xmp(f)[[1]]
     expect_snapshot(print(x2))
-    expect_equal(x$title, "An XMP title")
+    expect_equal(x$title[["x-default"]], "An XMP title")
 
     x <- xmp(creator = c("Creator 1", "Creator 2"))
     set_xmp(x, f)
@@ -63,20 +61,21 @@ test_that("get_xmp_exiftool() / set_xmp_exiftool()", {
     skip_if_not(supports_exiftool())
     f <- tempfile(fileext = ".pdf")
     on.exit(unlink(f))
-    pdf(f, onefile = TRUE)
+    pdf(f)
     grid.text("Page 1")
-    grid.newpage()
-    grid.text("Page 2")
     invisible(dev.off())
 
-    xmp <- xmp(title = "An XMP title", creator = "An XMP creator")
-    set_xmp_exiftool(xmp, f)
-    expect_equal(get_xmp_exiftool(f, use_names = FALSE)[[1]]$title, "An XMP title")
+    x <- xmp(creator = "An XMP creator")
+    x$title <- as_lang_alt(c("en" = "An English Title", "fr" = "A French Title"), default_lang = "en")
+    set_xmp_exiftool(x, f)
+    x <- get_xmp_exiftool(f, use_names = FALSE)[[1]]
+    expect_equal(x$title[["x-default"]], "An English Title")
+    expect_equal(x$title[["fr"]], "A French Title")
 })
 
 test_that("as_xmp()", {
     df <- data.frame(`dc.Creator` = "John Doe", `dc:Title` = "A Title",
                      check.names = FALSE, stringsAsFactors = FALSE)
     x <- as_xmp(df)
-    expect_equal(x$title, "A Title")
+    expect_equal(x$title[["x-default"]], "A Title")
 })